refactor: full reimplementation in TypeScript
This is a full reimplementation of the action in TypeScript, fully typed
and passing tests. Additionally, this implements retry logic from #329,
updates build dependencies slightly, and switches to the AWS SDK for
Javascript v3 (from v2).

The motivation for this was:
  * Increased readability and maintainability by splitting the project
    up into several files and type enforcement from the compiler.
  * Much smaller dist file by switching to JS v3's modular imports (now
    493 KB instead of 7.61 MB).
  * More realistic testing by using `aws-sdk-client-mock`

Admittedly, this is a large commit with a lot of changes that haven't
been previously discussed. No worries if a change of this scale isn't
accepted :)

closes: #299, #329, #321