feat: Better error message for GitHub OIDC
If a customer only configures a role and not accessKeyId
then assuming the customer intent is to use this role is reasonable.
If credentials could be derived in this case,
then we have introduced ambiguity into an authorization process.
And if credentials can not be derived,
then we have created a very confusing customer experience.