Remove assumptions about unhandled errors
The main function in index.js has a broad try/catch that handles all unhandled errors thrown when the action runs. However, it makes an intermittently incorrect assumption that the error thrown has a `.message` property. When this is not true, the stack trace is not emitted, preventing the end user from discovering the reason for the exception.

This commit changes the error handling to pass the entire error object to @ actions/toolkit/core, which then calls `.toString()` on the argument, effectively maintaining behavior for actual Error objects.

This commit does not correct the underlying causes of such errors; it only surfaces them